Cleanup some headers
